requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
python-multipart
trafilatura
newspaper4k
openai
python-docx
pydantic
httpx
